+++ server/models/dtos/validators.py
@@ -1,0 +1,10 @@
+from schematics.exceptions import ValidationError
+from server.models.statuses import ProjectStatus
+
+
+def is_known_project_status(value):
+    try:
+        ProjectStatus[value.upper()]
+    except KeyError:
+        raise ValidationError(f'Unknown projectStatus: {value} Valid values are {ProjectStatus.DRAFT.name}, '
+                              f'{ProjectStatus.PUBLISHED.name}, {ProjectStatus.ARCHIVED.name}')
